split map functions
